Moved VMState.get_parent_header(block_header, db) and VMState.get_prev_headers(last_block_hash, db) into standalone functions
