+++ evm/vm/forks/frontier/headers.py
@@ -12,7 +12,9 @@ from evm.constants import (
     BOMB_EXPONENTIAL_PERIOD,
     BOMB_EXPONENTIAL_FREE_PERIODS,
 )
-
+from evm.utils.db import (
+    get_parent_header,
+)
 from evm.utils.headers import (
     compute_gas_limit,
 )
@@ -90,7 +92,7 @@ def configure_frontier_header(vm, **header_params):
         setattr(vm.block.header, field_name, value)
 
     if 'timestamp' in header_params and vm.block.header.block_number > 0:
-        parent_header = vm.get_parent_header(vm.block.header, vm.chaindb)
+        parent_header = get_parent_header(vm.block.header, vm.chaindb)
         vm.block.header.difficulty = compute_frontier_difficulty(
             parent_header,
             header_params['timestamp'],
